+++ .github/workflows/test-with-optional-deps.yml
@@ -76,7 +76,7 @@ jobs:
           - "interop-crewai"
           - "interop-langchain"
           - "interop-pydantic-ai"
-          - "interop-mcp"
+          - "mcp"
           - "websockets"
           - "google-api"
         os: [ubuntu-latest, macos-latest, windows-latest]
