--- autogen/interop/mcp/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) 2023 - 2025, AG2ai, Inc., AG2ai open-source projects maintainers and core contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-
-from .mcp import MCPInteroperability
-
-__all__ = ["MCPInteroperability"]
