+++ autogen/tools/__init__.py
@@ -4,6 +4,7 @@
 
 from .dependency_injection import BaseContext, ChatContext, Depends
 from .function_utils import get_function_schema, load_basemodels_if_needed, serialize_to_str
+from .schema_defined_tool import SchemaDefinedTool
 from .tool import Tool, tool
 from .toolkit import Toolkit
 
@@ -11,6 +12,7 @@ __all__ = [
     "BaseContext",
     "ChatContext",
     "Depends",
+    "SchemaDefinedTool",
     "Tool",
     "Toolkit",
     "get_function_schema",
