+++ pyproject.toml
@@ -185,7 +185,7 @@ twilio = [
     "twilio>=9.3.2"
 ]
 
-interop-mcp = [
+mcp = [
     # mcp 1.6.0 keeps hanging in stdio_client context manager
     # test_mcp_issue_with_stdio_client_context_manager is failing for mcp=1.6.0
     "mcp>=1.4.0,<1.6; python_version>='3.10'"
@@ -198,7 +198,7 @@ interop-crewai = [
 interop-langchain = ["langchain-community>=0.3.12,<1"]
 interop-pydantic-ai = ["pydantic-ai==0.0.31"]
 interop =[
-    "pyautogen[interop-crewai, interop-langchain, interop-pydantic-ai, interop-mcp]",
+    "pyautogen[interop-crewai, interop-langchain, interop-pydantic-ai]"
 ]
 
 autobuild = ["chromadb", "sentence-transformers", "huggingface-hub"]
