+++ setup_ag2.py
@@ -46,7 +46,7 @@ setuptools.setup(
         "google-search": ["pyautogen[google-search]==" + __version__],
         "neo4j": ["pyautogen[neo4j]==" + __version__],
         "twilio": ["pyautogen[twilio]==" + __version__],
-        "interop-mcp": ["pyautogen[interop-mcp]==" + __version__],
+        "mcp": ["pyautogen[mcp]==" + __version__],
         "interop-crewai": ["pyautogen[interop-crewai]==" + __version__],
         "interop-langchain": ["pyautogen[interop-langchain]==" + __version__],
         "interop-pydantic-ai": ["pyautogen[interop-pydantic-ai]==" + __version__],
